Fix ARM build errors
Once I started trying to build the core for ARM, I encountered a few
bugs in the code which didn't show up in llvm used for OS X host
builds.
* I forgot to include some standard headers.
* I had to correct some casting so that it would work on 32-bit and
  64-bit architectures.

+++ Core/src/CrashCatcher.c
@@ -82,7 +82,7 @@ static uint32_t getAddressOfExceptionStack(const CrashCatcherExceptionRegisters*
 static const void* uint32AddressToPointer(uint32_t address)
 {
     if (sizeof(uint32_t*) == 8)
-        return (const void*)((uint64_t)address | g_crashCatcherTestBaseAddress);
+        return (const void*)(unsigned long)((uint64_t)address | g_crashCatcherTestBaseAddress);
     else
         return (const void*)(unsigned long)address;
 }
